# openapi.yaml
openapi: 3.0.3
info:
  title: Singapore Medical Portal API
  description: |
    API for medical examination portal supporting Singapore clinics with role-based access control,
    exam submission workflows, and approval management.
    
    **Supported Exam Types:**
    - Six-monthly Medical Exam for Migrant Domestic Workers (MOM)
    - Full Medical Exam for Work Permit (MOM)
    - Medical Exam for Aged Drivers (SPF)
    
    **User Roles:**
    - **Doctor**: Can create, approve, and submit exams directly to government
    - **Nurse**: Can create exams and route to doctors for approval
    - **Admin**: Can manage users and view all submissions
    
    **Workflow:**
    1. Nurse/Doctor creates exam (can save as draft)
    2. If nurse submits, exam goes to pending_approval status
    3. Doctor reviews and approves
    4. Approved exam is submitted to government portal
    
    **Security Note:** This is a demonstration API specification. Production implementation
    would require HIPAA-compliant infrastructure, proper encryption, audit logging, and
    secure authentication mechanisms.
    
  version: 1.0.0
  contact:
    name: CheckUp Medical Portal
    email: support@checkup.sg
  license:
    name: Proprietary
servers:
  - url: https://api.checkup.sg/v1
    description: Production server
  - url: https://staging-api.checkup.sg/v1
    description: Staging server
  - url: http://localhost:3000/v1
    description: Local development server
...
security:
  - bearerAuth: []
paths:
...
  /submissions:
    get:
      tags:
        - Submissions
      summary: List all submissions
      description: |
        Get list of medical exam submissions. Results are filtered based on user role:
        - Admins see all submissions
        - Doctors/Nurses see only their own submissions
      parameters:
        - name: status
          in: query
          description: Filter by submission status
          schema:
            type: string
            enum: [pending_approval, submitted, rejected]
        - name: examType
          in: query
          description: Filter by exam type
          schema:
            $ref: '#/components/schemas/ExamType'
        - name: patientName
          in: query
          description: Search by patient name (partial match)
          schema:
            type: string
        - name: patientNric
          in: query
          description: Search by patient NRIC/FIN
          schema:
            type: string
        - name: fromDate
          in: query
          description: Filter submissions from this date
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          description: Filter submissions until this date
          schema:
            type: string
            format: date
        - name: page
          in: query
          description: Page number for pagination
          schema:
            type: integer
            default: 1
            minimum: 1
        - name: limit
          in: query
          description: Number of items per page
          schema:
            type: integer
            default: 20
            minimum: 1
            maximum: 100
      responses:
        '200':
          description: List of submissions
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/MedicalSubmission'
                  pagination:
                    $ref: '#/components/schemas/Pagination'
        '401':
          $ref: '#/components/responses/UnauthorizedError'

    post:
      tags:
        - Submissions
      summary: Create new submission
      description: |
        Create a new medical exam submission. Behavior varies by role:
        - Doctors: Submission goes directly to 'submitted' status
        - Nurses: Submission goes to 'pending_approval' status unless routeForApproval is false
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - examType
                - patientName
                - patientNric
                - patientDateOfBirth
                - formData
              properties:
                examType:
                  $ref: '#/components/schemas/ExamType'
                patientName:
                  type: string
                  example: Maria Santos
                patientNric:
                  type: string
                  description: Singapore NRIC or FIN number
                  example: S1234567A
                patientDateOfBirth:
                  type: string
                  format: date
                  example: "1990-05-15"
                formData:
                  type: object
                  description: Exam-specific form fields
                  additionalProperties: true
                routeForApproval:
                  type: boolean
                  description: If true, route to doctor for approval (nurse only)
                  default: true
      responses:
        '201':
          description: Submission created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalSubmission'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /submissions/{id}:
    get:
      tags:
        - Submissions
      summary: Get submission by ID
      description: Retrieve detailed information about a specific submission
      parameters:
        - name: id
          in: path
          required: true
          description: Submission ID
          schema:
            type: string
            example: sub-001
      responses:
        '200':
          description: Submission details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalSubmission'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'

    put:
      tags:
        - Submissions
      summary: Update submission
      description: Update an existing submission (only if not yet submitted)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                patientName:
                  type: string
                patientNric:
                  type: string
                patientDateOfBirth:
                  type: string
                  format: date
                formData:
                  type: object
                  additionalProperties: true
      responses:
        '200':
          description: Submission updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalSubmission'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /submissions/{id}/submit:
    post:
      tags:
        - Submissions
      summary: Submit draft for approval
      description: |
        Submit a draft medical exam for approval by a doctor. 
        Only nurses can use this endpoint to route drafts to pending_approval status.
        Changes the status from 'draft' to 'pending_approval'.
      parameters:
        - name: id
          in: path
          required: true
          description: Submission ID
          schema:
            type: string
            example: sub-001
      responses:
        '200':
          description: Successfully submitted for approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MedicalSubmission'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          description: Forbidden - only drafts can be submitted for approval
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /submissions/{id}/history:
    get:
      tags:
        - Submissions
      summary: Get submission audit history
      description: |
        Retrieve the audit trail for a specific submission, including all status changes,
        updates, and approval actions.
      parameters:
        - name: id
          in: path
          required: true
          description: Submission ID
          schema:
            type: string
            example: sub-001
      responses:
        '200':
          description: Audit history retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  submissionId:
                    type: string
                  events:
                    type: array
                    items:
                      type: object
                      properties:
                        timestamp:
                          type: string
                          format: date-time
                        eventType:
                          type: string
                          enum: [created, updated, submitted, approved, rejected, deleted]
                        userId:
                          type: string
                        userName:
                          type: string
                        details:
                          type: object
        '404':
          $ref: '#/components/responses/NotFoundError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  schemas:
...
    ExamType:
      type: string
      enum:
        - Six-monthly Medical Exam for Migrant Domestic Workers (MOM)
        - Full Medical Exam for Work Permit (MOM)
        - Medical Exam for Aged Drivers (SPF)
      description: |
        Type of medical examination:
        - Six-monthly Medical Exam for Migrant Domestic Workers: Regular health screening for MDWs
        - Full Medical Exam for Work Permit: Comprehensive exam for work permit applications
        - Medical Exam for Aged Drivers: Medical assessment for elderly drivers
    SubmissionStatus:
      type: string
      enum: [draft, pending_approval, submitted, rejected]
      description: |
        - draft: Incomplete exam saved for later
        - pending_approval: Submitted by nurse, awaiting doctor approval
        - submitted: Approved and submitted to government portal
        - rejected: Rejected by doctor, requires revision
    MedicalSubmission:
      type: object
      required:
        - id
        - examType
        - patientName
        - patientNric
        - patientDateOfBirth
        - status
        - createdBy
        - createdByName
        - createdDate
        - clinicId
        - formData
      properties:
        id:
          type: string
          example: sub-001
        examType:
          $ref: '#/components/schemas/ExamType'
        patientName:
          type: string
          example: Maria Santos
        patientNric:
          type: string
          description: Singapore NRIC or FIN number
          example: S1234567A
        patientDateOfBirth:
          type: string
          format: date
          example: "1990-05-15"
        status:
          $ref: '#/components/schemas/SubmissionStatus'
        createdBy:
          type: string
          description: User ID who created the submission
          example: "2"
        createdByName:
          type: string
          example: Nurse Mary Lim
        createdDate:
          type: string
          format: date-time
          example: "2025-10-15T10:30:00Z"
        submittedDate:
          type: string
          format: date-time
          description: Date when submitted to government portal
          example: "2025-10-15T14:20:00Z"
        approvedBy:
          type: string
          description: Doctor ID who approved
          example: "1"
        approvedByName:
          type: string
          example: Dr. Sarah Tan
        approvedDate:
          type: string
          format: date-time
          example: "2025-10-15T14:00:00Z"
        rejectedReason:
          type: string
          description: Reason for rejection (if status is rejected)
        clinicId:
          type: string
          example: clinic1
        formData:
          type: object
          description: Exam-specific form fields
          additionalProperties: true
          example:
            height: "160"
            weight: "55"
            bloodPressure: "120/80"
            pregnancyTest: "Negative"
            chestXray: "Normal"
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 20
        totalPages:
          type: integer
          example: 5
        totalItems:
          type: integer
          example: 94
        hasNext:
          type: boolean
          example: true
        hasPrevious:
          type: boolean
          example: false
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: Invalid credentials
        code:
          type: string
          description: Error code
          example: AUTH_FAILED
        details:
          type: object
          additionalProperties: true
          description: Additional error details
  responses:
    UnauthorizedError:
      description: Authentication failed or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized
            code: UNAUTHORIZED
    ForbiddenError:
      description: User doesn't have permission for this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden - insufficient permissions
            code: FORBIDDEN
    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Resource not found
            code: NOT_FOUND
    BadRequestError:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Validation failed
            code: VALIDATION_ERROR
            details:
              patientNric: Invalid NRIC format